WIP converting to typescript
some issues I'm noticing

fetchLocalModel can sometimes return undefined, this is kind of annoying
it would be nice if it just created a new empty model if it can't find
the model.

If that's not possible we need to just create a new one in this
component I guess.

I defined the interface for PromoCodeAlert in two places, not sure where
I can define that in one place.